Extract Superduper mode
