+++ keyboards/planck/keymaps/narze/keymap.c
@@ -1,9 +1,5 @@
-// This is the canonical layout file for the Quantum project. If you want to add another keyboard,
-// this is the style you want to emulate.
-
 #include QMK_KEYBOARD_H
 #include "narze.h"
-#include "action_layer.h"
 #ifdef AUDIO_ENABLE
   #include "audio.h"
 #endif
@@ -50,30 +46,9 @@ enum planck_keycodes {
 #define SFT_ENT SFT_T(KC_ENT)
 #define GUI_MINS GUI_T(KC_MINS)
 
-// Combo : SuperDuper layer from S+D (R+S in Colemak)
-#define SUPERDUPER_COMBO_COUNT 3
-#define EECONFIG_SUPERDUPER_INDEX (uint8_t *) 19
-
 enum process_combo_event {
   CB_SUPERDUPER,
 };
-
-const uint16_t PROGMEM superduper_combos[SUPERDUPER_COMBO_COUNT][3] = {
-  [_QWERTY] = {KC_S, KC_D, COMBO_END},
-  [_COLEMAK] = {KC_R, KC_S, COMBO_END},
-  [_QWOC] = {CM_S, CM_D, COMBO_END},
-};
-
-combo_t key_combos[COMBO_COUNT] = {
-  [CB_SUPERDUPER] = COMBO_ACTION(superduper_combos[_QWERTY]),
-};
-
-volatile bool superduper_enabled = true;
-
-const uint16_t empty_combo[] = {COMBO_END};
-
-void set_superduper_key_combos(void);
-void clear_superduper_key_combos(void);
 
 const uint16_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
 
@@ -261,8 +236,7 @@ bool process_record_user(uint16_t keycode, keyrecord_t *record) {
         #endif
         persistant_default_layer_set(1UL<<_QWERTY);
 
-        key_combos[CB_SUPERDUPER].keys = superduper_combos[_QWERTY];
-        eeprom_update_byte(EECONFIG_SUPERDUPER_INDEX, _QWERTY);
+        set_superduper_key_combo_layer(_QWERTY);
       }
       return false;
       break;
@@ -273,8 +247,7 @@ bool process_record_user(uint16_t keycode, keyrecord_t *record) {
         #endif
         persistant_default_layer_set(1UL<<_COLEMAK);
 
-        key_combos[CB_SUPERDUPER].keys = superduper_combos[_COLEMAK];
-        eeprom_update_byte(EECONFIG_SUPERDUPER_INDEX, _COLEMAK);
+        set_superduper_key_combo_layer(_COLEMAK);
       }
       return false;
       break;
@@ -285,8 +258,7 @@ bool process_record_user(uint16_t keycode, keyrecord_t *record) {
         #endif
         persistant_default_layer_set(1UL<<_QWOC);
 
-        key_combos[CB_SUPERDUPER].keys = superduper_combos[_QWOC];
-        eeprom_update_byte(EECONFIG_SUPERDUPER_INDEX, _QWOC);
+        set_superduper_key_combo_layer(_QWOC);
       }
       return false;
       break;
@@ -347,21 +319,7 @@ bool process_record_user(uint16_t keycode, keyrecord_t *record) {
       return false;
     case SDTOGG:
       if (record->event.pressed) {
-        superduper_enabled = !superduper_enabled;
-
-        if (superduper_enabled) {
-          set_superduper_key_combos();
-
-          #ifdef AUDIO_ENABLE
-            PLAY_NOTE_ARRAY(tone_sonic_ring, false, 0);
-          #endif
-        } else {
-          clear_superduper_key_combos();
-
-          #ifdef AUDIO_ENABLE
-            PLAY_NOTE_ARRAY(tone_coin, false, 0);
-          #endif
-        }
+        toggle_superduper_mode();
       }
       return false;
 
@@ -393,22 +351,6 @@ void matrix_setup(void) {
   set_superduper_key_combos();
 }
 
-void set_superduper_key_combos(void) {
-  uint8_t layer = eeprom_read_byte(EECONFIG_SUPERDUPER_INDEX);
-
-  switch (layer) {
-    case _QWERTY:
-    case _COLEMAK:
-    case _QWOC:
-      key_combos[CB_SUPERDUPER].keys = superduper_combos[layer];
-      break;
-  }
-}
-
-void clear_superduper_key_combos(void) {
-  key_combos[CB_SUPERDUPER].keys = empty_combo;
-}
-
 void matrix_scan_user(void) {
 }
 
@@ -438,8 +380,6 @@ void music_scale_user(void)
 }
 
 #endif
-
-// Combos
 
 void process_combo_event(uint8_t combo_index, bool pressed) {
   if (pressed) {
